Added timer, also made the GUI into a separate class to keep Main clean
