+++ OOP19_Group5/src/application/MainGUI.java
@@ -1,0 +1,171 @@
+package application;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.geometry.Insets;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
+import javafx.scene.control.TextField;
+import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.VBox;
+import javafx.stage.Stage;
+
+public class MainGUI {
+
+	TableView <Skier> table;
+	TextField nameInput, startNumberInput;
+	boolean running;
+
+	Label lblDifference = new Label();
+	AniTimer timer = new AniTimer(lblDifference);
+
+	public MainGUI() {
+		
+	}
+	
+	
+	public void display() {
+		Stage primaryStage = new Stage();
+		
+		try {
+			/** Kolumner **/
+			// NamnKolumn
+			TableColumn<Skier, String> fullNameColumn = new TableColumn<>("Name");
+			fullNameColumn.setMinWidth(170);
+			fullNameColumn.setCellValueFactory(new PropertyValueFactory<>("fullName"));
+
+			// Startnummer
+			TableColumn<Skier, Integer> startNumberColumn = new TableColumn<>("Startnumber");
+			startNumberColumn.setMinWidth(100);
+			startNumberColumn.setCellValueFactory(new PropertyValueFactory<>("startNumber"));
+
+			// Differens
+			TableColumn<Skier, Integer> differenceColumn = new TableColumn<>("Difference");
+			differenceColumn.setMinWidth(170);
+			differenceColumn.setCellValueFactory(new PropertyValueFactory<>("startNumber"));
+
+			table = new TableView<>();
+			table.setItems(getSkier());
+			table.getColumns().addAll(fullNameColumn, startNumberColumn, differenceColumn);
+
+			nameInput = new TextField();
+			nameInput.setPromptText("Name");
+			nameInput.setMaxWidth(170);
+
+			startNumberInput = new TextField();
+			startNumberInput.setPromptText("Startnumber");
+			startNumberInput.setMaxWidth(170);
+
+			// Knappar
+			Button btnSearchName = new Button("Search");
+			Button btnSearchStartNumber = new Button("Search");
+			Button btnUpdate = new Button("Add and refresh list");
+			Button btnStartRace = new Button("Start race");
+
+			btnStartRace.setOnAction(e->{
+				if(!running) {
+					timer.start();
+					btnStartRace.setText("Stop race");
+					running = true;
+				}
+				else {
+					timer.stop();
+					timer.reset();
+					btnStartRace.setText("Start race");
+					running = false;
+
+				}
+
+
+
+			});
+			Button btnHunt = new Button("Hunt");
+			Button btnIndividual15 = new Button("Mass");
+			Button btnIndividual30 = new Button("30 sek");
+
+			// Layout
+			BorderPane root = new BorderPane();
+			VBox vBoxRight = new VBox();
+			GridPane gridPane = new GridPane();
+			gridPane.setPadding(new Insets(10, 10, 10, 10));
+			gridPane.setVgap(30  );
+			gridPane.setHgap(5);
+
+			// Name and time current skier
+			Label lblNameCurrentSkier = new Label();
+			lblNameCurrentSkier.setStyle("-fx-font-size: 18");
+			lblNameCurrentSkier.setText("Gunde Svan");
+			GridPane.setConstraints(lblNameCurrentSkier, 0, 0);
+			Label lblTimeCurrentSkier = new Label();
+			lblTimeCurrentSkier.setStyle("-fx-font-size: 18");
+			lblTimeCurrentSkier.setText("00:00:00");
+
+			GridPane.setConstraints(lblTimeCurrentSkier, 1, 0);
+
+			// Search skier
+			GridPane.setConstraints(btnSearchName, 0, 1);
+			GridPane.setConstraints(btnSearchStartNumber, 0, 2);
+			GridPane.setConstraints(nameInput, 1, 1);
+			GridPane.setConstraints(startNumberInput, 1, 2);
+
+			// Update
+			GridPane.setConstraints(btnUpdate, 0, 3);
+
+			// Show leader
+			Label lblNrCurrentLeader = new Label();
+			lblNrCurrentLeader.setStyle("-fx-font-size: 20");
+			lblNrCurrentLeader.setText("Lucas Bauer");
+			GridPane.setConstraints(lblNrCurrentLeader, 0, 4);
+
+			// Show difference to leader
+
+			lblDifference.setStyle("-fx-font-size: 40");
+			lblDifference.setText("00:00:00");
+			GridPane.setConstraints(lblDifference, 0, 5);
+
+			// RaceSettings
+			GridPane.setConstraints(btnStartRace, 0, 6);
+			Label lblTypeOfRace = new Label("Select type of start");
+			GridPane.setConstraints(lblTypeOfRace, 0, 7);
+
+
+			gridPane.getChildren().addAll(lblNameCurrentSkier, lblTimeCurrentSkier, btnSearchName, nameInput, btnSearchStartNumber, startNumberInput,
+					btnUpdate, lblNrCurrentLeader, lblDifference, btnStartRace, lblTypeOfRace);
+
+			HBox hBox = new HBox();
+			hBox.setPadding(new Insets(10, 10, 10, 10));
+			hBox.setSpacing(20);
+			hBox.getChildren().addAll(btnHunt, btnIndividual15, btnIndividual30);
+
+			vBoxRight.setPadding(new Insets(10, 10, 10,10));
+			vBoxRight.getChildren().addAll(gridPane, hBox);
+
+			root.setLeft(table);
+			root.setRight(vBoxRight);
+
+			Scene scene = new Scene(root,800,900);
+			scene.getStylesheets().add(getClass().getResource("application.css").toExternalForm());
+			primaryStage.setScene(scene);
+			primaryStage.setTitle("Russian EPOGenerator");
+			primaryStage.show();
+		} catch(Exception e) {
+			e.printStackTrace();
+		}
+
+	}
+
+	public ObservableList<Skier> getSkier() {
+		ObservableList<Skier> skiers = FXCollections.observableArrayList();
+		skiers.add(new Skier("Gunde", "Svan", 1));
+		skiers.add(new Skier("Dimitri", "Youshenko", 2));
+		skiers.add(new Skier("Lucas", "Bauer", 3));
+		return skiers;
+	}
+
+}
